Add go version workflow

+++ gofsutil_fs.go
@@ -192,15 +192,21 @@ func (fs *FS) fsInfo(_ context.Context, path string) (int64, int64, int64, int64
 	}
 
 	// Available is blocks available * fragment size
-	available := int64(statfs.Bavail) * int64(statfs.Bsize)
+	// #nosec G115
+	available := int64(statfs.Bavail) * statfs.Bsize
 
 	// Capacity is total block count * fragment size
-	capacity := int64(statfs.Blocks) * int64(statfs.Bsize)
+	// #nosec G115
+	capacity := int64(statfs.Blocks) * statfs.Bsize
 
 	// Usage is block being used * fragment size (aka block size).
-	usage := (int64(statfs.Blocks) - int64(statfs.Bfree)) * int64(statfs.Bsize)
+	// #nosec G115
+	usage := (int64(statfs.Blocks) - int64(statfs.Bfree)) * statfs.Bsize
 
+	// #nosec G115
 	inodes := int64(statfs.Files)
+
+	// #nosec G115
 	inodesFree := int64(statfs.Ffree)
 	inodesUsed := inodes - inodesFree
